+++ include/poac/core/resolver.hpp
@@ -37,7 +37,6 @@ using IncorrectSha256sum = Error<
     "issue on GitHub of the package and stopping using this package:\n"
     "  {}: {}",
     String, String, String, String>;
-using ConanNotFound = Error<"conan is not found">;
 using Unknown = Error<"unknown error occurred: {}", String>;
 
 inline Fn get_install_name(const resolve::Package& package)->String {
