+++ include/poac/core/resolver/resolve.hpp
@@ -35,8 +35,6 @@ struct DependencyInfo {
   ///
   /// Sometimes, this is like `1.66.0` or like `>=1.64.0 and <2.0.0`.
   String version_rq;
-
-  bool is_conan() const noexcept { return name.find("conan::") == 0; }
 };
 
 inline Fn operator==(const DependencyInfo& lhs, const DependencyInfo& rhs)
@@ -56,6 +54,8 @@ struct Package {
   String name;
 
   DependencyInfo dep_info;
+
+  bool is_conan() const noexcept { return name.find("conan::") == 0; }
 };
 
 inline Fn operator==(const Package& lhs, const Package& rhs)->bool {
