+++ lib/core/builder/CMakeLists.txt
@@ -19,4 +19,5 @@ add_poac_target(poac_core_builder
         poac_util_log
         poac_util_result
         poac_util_rustify
+        poac_util_registry_conan_manifest
 )
